+++ src/typography/fonts/DMMono/index.ts
@@ -1,0 +1,1 @@
+export * from "./DMMono.font";