+++ src/typography/fonts/index.ts
@@ -1,1 +1,2 @@
 export * from "./Gilroy";
+export * from "./DMMono";