+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Update styles while hovering, active & disabled
   - Improve accessibity by setting the button's title dynamically
 
+### Added
+
+- New font: DM Mono
+- Support for it in all typography components
+
 ## 1.8.15 - 2025-04-10
 ### Fixed
 - `v0.0.5` Hypertext component